#103 [move - user] Enum class 패키지 이동
